requirements-fastapi.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
psycopg2-binary
